requirements.txt
fastmcp>=0.1.0
lmstudio
pymongo>=4.0.0
paho-mqtt>=2.0.0
python-dotenv>=0.19.0
uvicorn>=0.17.0
starlette>=0.17.1
scikit-learn>=0.0
numpy>=1.20.0
python-dateutil>=2.8.2
python-jose>=3.3.0
